added flash and checkbox components

+++ app/javascript/components/Application.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import TopNav from "./shared/TopNav";
+import Flash from "./shared/Flash";
 import UsersDashBoard from "./UsersDashBoard";
 import GroupsDashBoard from "./GroupsDashBoard";
 import { Route, Redirect, Switch } from "react-router-dom";
@@ -7,7 +8,8 @@ import { Route, Redirect, Switch } from "react-router-dom";
 class Application extends React.Component {
   state = {
     users: [],
-    groups: []
+    groups: [],
+    flash: { message: "", type: "" }
   };
 
   componentDidMount() {
@@ -46,6 +48,7 @@ class Application extends React.Component {
       })
       .then(user => {
         this.addNewUser(user);
+        this.flashMessage("A new user has been added successfully", "success");
         return fetch("api/groups");
       })
       .then(response => {
@@ -54,6 +57,18 @@ class Application extends React.Component {
       .then(json => {
         this.setState({ groups: json });
       });
+  };
+
+  flashMessage = (message, type) => {
+    this.setState({
+      flash: {
+        message: message,
+        type: type
+      }
+    });
+    setTimeout(() => {
+      this.resetFlashState();
+    }, 3000);
   };
 
   handleCreateGroup = (name, description) => {
@@ -78,6 +93,7 @@ class Application extends React.Component {
       })
       .then(group => {
         this.addNewGroup(group);
+        this.flashMessage("A new group was successfully added", "success");
       });
   };
 
@@ -94,14 +110,22 @@ class Application extends React.Component {
   };
 
   handleDeleteUser = id => {
-    fetch(`/api/users/${id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json"
-      }
-    }).then(response => {
-      this.deleteUser(id);
-    });
+    if (this.state.users.find(user => user.id === id).groups.length === 0) {
+      fetch(`/api/users/${id}`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json"
+        }
+      }).then(response => {
+        this.deleteUser(id);
+        this.flashMessage("User was successfully deleted", "success");
+      });
+    } else {
+      this.flashMessage(
+        "Cannot delete user that is still subscribed to groups",
+        "error"
+      );
+    }
   };
 
   deleteUser = id => {
@@ -112,22 +136,30 @@ class Application extends React.Component {
   };
 
   handleDeleteGroup = id => {
-    fetch(`/api/groups/${id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json"
-      }
-    })
-      .then(response => {
-        this.deleteGroup(id);
-        return fetch("api/users");
+    if (this.state.groups.find(group => group.id === id).users.length === 0) {
+      fetch(`/api/groups/${id}`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json"
+        }
       })
-      .then(response => {
-        return response.json();
-      })
-      .then(json => {
-        this.setState({ users: json });
-      });
+        .then(response => {
+          this.deleteGroup(id);
+          return fetch("api/users");
+        })
+        .then(response => {
+          return response.json();
+        })
+        .then(json => {
+          this.setState({ users: json });
+          this.flashMessage("Group was successfully deleted", "success");
+        });
+    } else {
+      this.flashMessage(
+        "Cannot delete a group that still has members",
+        "error"
+      );
+    }
   };
 
   deleteGroup = id => {
@@ -154,6 +186,7 @@ class Application extends React.Component {
       })
       .then(user => {
         this.updateUser(user);
+        this.flashMessage("User has been successfully edited", "success");
         return fetch("api/groups");
       })
       .then(response => {
@@ -184,6 +217,7 @@ class Application extends React.Component {
       })
       .then(group => {
         this.updateGroup(group);
+        this.flashMessage("Group has been successfully edited", "success");
       });
   };
 
@@ -214,6 +248,10 @@ class Application extends React.Component {
     });
   };
 
+  resetFlashState = () => {
+    this.setState({ flash: { message: "", type: "" } });
+  };
+
   render() {
     const { users, groups } = this.state;
     const groupNames = groups.map(group => group.name);
@@ -221,6 +259,7 @@ class Application extends React.Component {
     return (
       <div className="container">
         <TopNav />
+        <Flash flash={this.state.flash} onFlashClose={this.resetFlashState} />
         <Switch>
           <Route
             path="/users"
